Fix TODO strange callback bug with lambdas.
The issue caused all interactable elements in the Menu CTKFrame
instances to only execute the command of the last element.

This was an issue with scope, and using a variable that iterates in the
scope of a for loop. That varible is as a key to execute the correct
command for the interactable element.

A default assignment was used to create a varible in the lambda scope,
that is then refered to as the key.

See: https://stackoverflow.com/questions/7546285/creating-lambda-inside-a-loop