Fix no attribute error and rename some variables and functions.
This commit fixes an issue where using a keybind to close the
application in a window manager causes a no attribute error, resulting
in the application being frozen in a blank state. The duplicate
Notification instances now initializes the self.selected_spot,
referenced in the destroy() function.

It also removes a TODO, which seems to be implemented by the thing that
causes this issue.